Split test_asttools to reflect split in asttools
